Adds optional description to `ConfigEntry`
The main reason to do this is to be able to use the
class `ConfigEntry` as the source of documentation
for your project.

Similar to what `--help` does for command line arguments,
you could create all the config entries and output
the expected configuration in your program (or use it
to generate readmes or so).

Changes added:
- New public attribute of type `Option[String]`
- New method `withDescription` to "set" its value

If this idea looks good to the maintainers of the repo
we could extend it a bit further by adding convenience
methods to set the description in `env`, `arg` and the likes.

Signed-off-by: Albert Pastrana <albert.pastrana@intenthq.com>